Fix the bug about comparing with mismatched integrate type
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@5147 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/MdePkg/Include/Library/HobLib.h
@@ -393,7 +393,7 @@ BuildMemoryAllocationHob (
 #define GET_HOB_TYPE(Hob)     ((Hob).Header->HobType)
 #define GET_HOB_LENGTH(Hob)   ((Hob).Header->HobLength)
 #define GET_NEXT_HOB(Hob)     ((Hob).Raw + GET_HOB_LENGTH (Hob))
-#define END_OF_HOB_LIST(Hob)  (GET_HOB_TYPE (Hob) == EFI_HOB_TYPE_END_OF_HOB_LIST)
+#define END_OF_HOB_LIST(Hob)  (GET_HOB_TYPE (Hob) == (UINT16)EFI_HOB_TYPE_END_OF_HOB_LIST)
 
 //
 // Get the data and data size field of GUID
